requirements.txt
google-genai
python-dotenv
streamlit
yt-dlp